Generate binary treatment variable

The proportion of assignment to treatment is p_trt. Generation can be done fixing the proportion of patients in the treatment group to be as close as possible to p_trt and then permuting (type = "exact") or random sampling from a binomial distribution with probability p_trt (type = "random"). For type = "random" one can also specify a formula for logit(P(trt|X)) via "prop". The probability of obtaining treatment is then binomial with P(trt|X) = 1/(1+exp(-prop)). The variables names specified in "prop" need to be handed over in a data frame "X".

generate_trt(n, p_trt = 0.5, type = c("exact", "random"), X = NULL,
  prop = NULL)

Arguments

n
Total sample size

p_trt
Proportion of treatment assignment (default = 0.5)

type
Either "exact" (proportion of p_trt is tried to match exactly), or "random" then treatment assignment is generated from a binomial distribution.

X
matrix/dataframe of predictor variables

prop
A formula used to generate logit(P(trt|X)). Probability of obtaining treatment is then Bin(1, P(trt|X)) with P(trt|X) = 1/(1+exp(-prop)). If type = "random" and prop is not specified p_trt is used in the binomial distribution

Value

Vector of treatment indicators

Examples

## by default try to match target trt proportion in observed trt proportion
trt <- generate_trt(10) # p_trt = 0.5 is default
table(trt)

#> trt
#> 0 1 
#> 5 5 

trt <- generate_trt(10, p_trt = 0.2)
trt <- generate_trt(11, p_trt = 0.5)
table(trt) # remaining patient randomly allocated

#> trt
#> 0 1 
#> 5 6 

## example for random allocation
trt <- generate_trt(10, p_trt = 0.5, type = "random") # samples from binomial with p_trt = 0.5
table(trt) # observed allocation may deviate from p_trt=0.5 for small samples

#> trt
#> 0 1 
#> 7 3 

## example, where treatment allocation depends on propensity model
X <- generate_X_dist(n=10000, p=10, rho=0.5)
## Use propensity score for treatment as 1/(1+exp(-X2))
trt <- generate_trt(n=nrow(X), type = "random", X=X, prop = "X2")
## fit propensity model
dat <- cbind(trt, X)
fit <- glm(trt~., data=dat, family=binomial)
summary(fit)

#> 
#> Call:
#> glm(formula = trt ~ ., family = binomial, data = dat)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -2.7874  -1.2310   0.6220   0.9384   1.2743  
#> 
#> Coefficients:
#>              Estimate Std. Error z value Pr(>|z|)    
#> (Intercept) -0.073530   0.062417  -1.178   0.2388    
#> X1Y          0.069048   0.045561   1.515   0.1296    
#> X2           0.983315   0.036969  26.598   <2e-16 ***
#> X3          -0.052008   0.026948  -1.930   0.0536 .  
#> X4           0.006089   0.026803   0.227   0.8203    
#> X5Y          0.057717   0.051690   1.117   0.2642    
#> X6M2        -0.036260   0.063957  -0.567   0.5708    
#> X6M3         0.075191   0.064340   1.169   0.2425    
#> X6M4         0.047773   0.064442   0.741   0.4585    
#> X7           0.037181   0.023001   1.617   0.1060    
#> X8          -0.027878   0.022662  -1.230   0.2186    
#> X9           0.049363   0.023003   2.146   0.0319 *  
#> X10          0.008969   0.022470   0.399   0.6898    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 12330  on 9999  degrees of freedom
#> Residual deviance: 11266  on 9987  degrees of freedom
#> AIC: 11292
#> 
#> Number of Fisher Scoring iterations: 5
#>