Generate binary treatment variable

The proportion of assignment to treatment is p_trt. Generation can be done fixing the proportion of patients in the treatment group to be as close as possible to p_trt and then permuting (type = "exact") or random sampling from a binomial distribution with probability p_trt (type = "random"). For type = "random" one can also specify a formula for logit(P(trt|X)) via "prop". The probability of obtaining treatment is then binomial with P(trt|X) = 1/(1+exp(-prop)). The variables names specified in "prop" need to be handed over in a data frame "X".

generate_trt(
  n,
  p_trt = 0.5,
  type = c("exact", "random"),
  X = NULL,
  prop = NULL
)

Arguments

n

Total sample size

p_trt

Proportion of treatment assignment (default = 0.5)

type

Either "exact" (proportion of p_trt is tried to match exactly), or "random" then treatment assignment is generated from a binomial distribution.

X

matrix/dataframe of predictor variables

prop

A formula used to generate logit(P(trt|X)). Probability of obtaining treatment is then Bin(1, P(trt|X)) with P(trt|X) = 1/(1+exp(-prop)). If type = "random" and prop is not specified p_trt is used in the binomial distribution

Value

Vector of treatment indicators

Examples

## by default try to match target trt proportion in observed trt proportion
trt <- generate_trt(10) # p_trt = 0.5 is default
table(trt)
#> trt
#> 0 1 
#> 5 5 
trt <- generate_trt(10, p_trt = 0.2)
trt <- generate_trt(11, p_trt = 0.5)
table(trt) # remaining patient randomly allocated
#> trt
#> 0 1 
#> 5 6 
## example for random allocation
trt <- generate_trt(10, p_trt = 0.5, type = "random") # samples from binomial with p_trt = 0.5
table(trt) # observed allocation may deviate from p_trt=0.5 for small samples
#> trt
#> 0 1 
#> 6 4 

## example, where treatment allocation depends on propensity model
X <- generate_X_dist(n=10000, p=10, rho=0.5)
## Use propensity score for treatment as 1/(1+exp(-X2))
trt <- generate_trt(n=nrow(X), type = "random", X=X, prop = "X2")
## fit propensity model
dat <- cbind(trt, X)
fit <- glm(trt~., data=dat, family=binomial)
summary(fit)
#> 
#> Call:
#> glm(formula = trt ~ ., family = binomial, data = dat)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -3.0089  -1.2140   0.6143   0.9556   1.2836  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept) -0.04528    0.06105  -0.742   0.4583    
#> X1Y         -0.00191    0.04523  -0.042   0.9663    
#> X2           1.01391    0.03715  27.291   <2e-16 ***
#> X3           0.03654    0.02711   1.348   0.1777    
#> X4          -0.03515    0.02718  -1.294   0.1958    
#> X5Y          0.01131    0.05050   0.224   0.8227    
#> X6M2        -0.03286    0.06385  -0.515   0.6068    
#> X6M3         0.02525    0.06385   0.395   0.6926    
#> X6M4        -0.06203    0.06338  -0.979   0.3277    
#> X7          -0.01809    0.02282  -0.793   0.4279    
#> X8          -0.01183    0.02281  -0.519   0.6039    
#> X9          -0.04832    0.02255  -2.143   0.0322 *  
#> X10         -0.04983    0.02273  -2.192   0.0284 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 12506  on 9999  degrees of freedom
#> Residual deviance: 11391  on 9987  degrees of freedom
#> AIC: 11417
#> 
#> Number of Fisher Scoring iterations: 5
#>