Generate binary treatment variable

The proportion of assignment to treatment is p_trt. Generation can be done fixing the proportion of patients in the treatment group to be as close as possible to p_trt and then permuting (type = "exact") or random sampling from a binomial distribution with probability p_trt (type = "random"). For type = "random" one can also specify a formula for logit(P(trt|X)) via "prop". The probability of obtaining treatment is then binomial with P(trt|X) = 1/(1+exp(-prop)). The variables names specified in "prop" need to be handed over in a data frame "X".

generate_trt(
  n,
  p_trt = 0.5,
  type = c("exact", "random"),
  X = NULL,
  prop = NULL
)

Arguments

n

Total sample size

p_trt

Proportion of treatment assignment (default = 0.5)

type

Either "exact" (proportion of p_trt is tried to match exactly), or "random" then treatment assignment is generated from a binomial distribution.

X

matrix/dataframe of predictor variables

prop

A formula used to generate logit(P(trt|X)). Probability of obtaining treatment is then Bin(1, P(trt|X)) with P(trt|X) = 1/(1+exp(-prop)). If type = "random" and prop is not specified p_trt is used in the binomial distribution

Value

Vector of treatment indicators

Examples

## by default try to match target trt proportion in observed trt proportion
trt <- generate_trt(10) # p_trt = 0.5 is default
table(trt)
#> trt
#> 0 1 
#> 5 5 
trt <- generate_trt(10, p_trt = 0.2)
trt <- generate_trt(11, p_trt = 0.5)
table(trt) # remaining patient randomly allocated
#> trt
#> 0 1 
#> 5 6 
## example for random allocation
trt <- generate_trt(10, p_trt = 0.5, type = "random") # samples from binomial with p_trt = 0.5
table(trt) # observed allocation may deviate from p_trt=0.5 for small samples
#> trt
#> 0 1 
#> 6 4 

## example, where treatment allocation depends on propensity model
X <- generate_X_dist(n=10000, p=10, rho=0.5)
## Use propensity score for treatment as 1/(1+exp(-X2))
trt <- generate_trt(n=nrow(X), type = "random", X=X, prop = "X2")
## fit propensity model
dat <- cbind(trt, X)
fit <- glm(trt~., data=dat, family=binomial)
summary(fit)
#> 
#> Call:
#> glm(formula = trt ~ ., family = binomial, data = dat)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -2.8812  -1.2201   0.6168   0.9519   1.2714  
#> 
#> Coefficients:
#>              Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)  0.057982   0.062975   0.921   0.3572    
#> X1Y         -0.036249   0.045378  -0.799   0.4244    
#> X2           1.015636   0.037495  27.087   <2e-16 ***
#> X3          -0.007566   0.027385  -0.276   0.7823    
#> X4           0.001991   0.026993   0.074   0.9412    
#> X5Y          0.022418   0.051468   0.436   0.6632    
#> X6M2        -0.108352   0.064535  -1.679   0.0932 .  
#> X6M3        -0.164570   0.064558  -2.549   0.0108 *  
#> X6M4        -0.060581   0.065195  -0.929   0.3528    
#> X7           0.024485   0.022575   1.085   0.2781    
#> X8           0.001304   0.022800   0.057   0.9544    
#> X9           0.048198   0.022705   2.123   0.0338 *  
#> X10          0.016837   0.022470   0.749   0.4537    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 12436  on 9999  degrees of freedom
#> Residual deviance: 11329  on 9987  degrees of freedom
#> AIC: 11355
#> 
#> Number of Fisher Scoring iterations: 5
#>